+++ Sources/BraintreeShopperInsights/Models/BTShopperInsightsCampaign.swift
@@ -1,0 +1,18 @@
+import Foundation
+
+/// A PayPal campaign displayed to the customer in the shopping journey.
+/// - Warning: This feature is in beta. It's public API may change or be removed in future releases.
+public struct BTShopperInsightsCampaign: Encodable {
+
+    // MARK: - Internal Properties
+
+    let id: String
+
+    // MARK: - Public Initializer
+
+    /// Creates a BTShopperInsightsCampaign.
+    /// - Parameter id: Required. The PayPal campaign ID.
+    public init(id: String) {
+        self.id = id
+    }
+}